+++ src/model/DateTime.js
@@ -1,4 +1,4 @@
-import { isString } from '../utils';
+import { isString } from '../util';
 
 class DateTime {
   constructor(rawObj) {
