+++ src/model/NcxItem.js
@@ -1,4 +1,4 @@
-import { mergeObjects } from '../utils';
+import { mergeObjects } from '../util';
 import Item from './Item';
 import NavPoint from './NavPoint';
