+++ test/util/mergeObjects.spec.js
@@ -1,0 +1,43 @@
+import { should } from 'chai';
+
+import mergeObjects from '../../src/util/mergeObjects';
+
+should(); // Initialize should
+
+describe('Util - Merge objects', () => {
+  it('mergeObjects test', () => {
+    const expect = {
+      a: true,
+      b: true,
+      c: {
+        d: 'd',
+        e: undefined,
+        f: {
+          g: 8,
+        },
+        h: '',
+      },
+    };
+    mergeObjects(
+      {
+        a: 'a',
+        b: true,
+        c: {
+          d: 5,
+          e: undefined,
+          f: {},
+        },
+      },
+      {
+        a: true,
+        c: {
+          d: 'd',
+          f: {
+            g: 8,
+          },
+          h: '',
+        },
+      },
+    ).should.deep.equal(expect);
+  });
+});
